# [FIX] l10n_in: improve perfs of l10n_in invoice report.