Slightly change the view of account_invoice_report to add
a CTE to prefilter rows before joining with account_move_line.

Add a multicolumn index for (product_id, move_id) on account_move_line
to speed-up cess_amount computation.

The speed-up depends on the the result of the CTE in l10_account_invoice_report sql view.
When there is no output rows from the CTE, the speed-up
for GSTR-1 reporting this month is from ~5m45 to ~35s, so close to 10.

If the filter does not apply, then the execution time goes to ~45s so closer to 8.

Some more speed-up data:
  - (with filter) ~10m to ~1.25m for GSTR-1 Reporting Last Year, speedup ~= 7.
  - (with filter) ~1m25s to ~8.5s for clicking on line in GSTR-1 Reporting this month view. speedup ~= 10.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr