@yelizariev 2021-07-09
For opw-2585805 this solves timeout problem, though still takes 12-14 minutes.

r+ ?