@yelizariev 2021-07-09
@jva-odoo would you like to review?