@jva-odoo 2021-07-09
@yelizariev Yes, plz