@jva-odoo 2021-07-09
@yelizariev it looks goods.