@jco-odoo 2021-07-12
@jva-odoo Can you make someone test on your side as well?