@jva-odoo 2021-07-12
@jco-odoo okay, I already test this and now I give this to someone functional guys.