@jva-odoo 2021-07-13
@Aurelienvd syntax error https://runbot.odoo.com/runbot/build/8867042