@Aurelienvd 2021-07-30
@odoo/rd-accounting I think this PR can be r+ (reviewed and tested by jva)