@Aurelienvd 2021-07-30
@jco-odoo I closed odoo/enterprise#16815 as the speedup wasn't consistent nor significant enough to justify the changes.